fix(core): uniform fp32 artifacts + all-or-nothing whitener gate everywhere
Closes two cross-surface consistency gaps found in a pre-release audit of
extraction / probing / steering.

All-or-nothing Mahalanobis gate (F1): `Profile.cosine_similarity` and
`project_profile` were the two surfaces the gate rollout missed — they used
a per-layer `whitener.covers()` check and could mix Mahalanobis and Euclidean
across layers (incommensurable `1/sqrt(lambda_L)` scales under cross-layer
normalization), which `vector compare`'s aggregate and Mahalanobis `~`/`|`
projection inherited. Both now decide the metric once via `covers_all` over
the relevant layer set (shared layers / `base ∩ onto`) and fall back to
all-Euclidean on partial coverage, matching monitor / bake / manifold.

Uniform fp32 storage (F2/F3): every saklas safetensor artifact is now stored
fp32, cast at the writer. The per-model neutral-activation cache was the last
exception (bf16) — that lost ~0.4% precision on exactly the quantity the
whitener inverts (`K=(NlambdaI+XX^T)^-1`) and created a seam where the
compute path (cache miss) returned fresh fp32 while the cache-hit / from_cache
paths returned bf16->fp32, so the whitener was not bit-reproducible across the
cache boundary. `load_or_compute_neutral_activations` now stores fp32 and
invalidates legacy bf16/fp16 caches (raw-on-disk dtype check -> recompute);
`save_profile` and `save_manifold` cast fp32 at write. fp32 keeps the
fp16-overflow fix (range) without bf16's input-precision loss.

from_cache hole: `LayerWhitener.from_cache` (the `vector compare` path) loaded
neutral activations directly, bypassing the invalidation. It has no model to
recompute with, so it now refuses a legacy non-fp32 cache with a repopulate
hint rather than silently building a reduced-precision metric.

Docs (F4): corrected the stale fp16/bf16 mentions and per-layer-fallback
wording across the AGENTS.md files, parsers help, and session docstring.

Tests: test_metric_gate_consistency.py (partial->all-Euclidean,
full->all-Mahalanobis on both surfaces) and test_neutral_acts_fp32.py (fp32
store, bit-identical compute/cache, bf16 invalidation, from_cache rejection).

+++ saklas/core/manifold.py
@@ -1504,17 +1504,19 @@ def save_manifold(
     path = Path(path)
     path.parent.mkdir(parents=True, exist_ok=True)
 
+    # Every saklas safetensor artifact is stored fp32 (cast at the writer);
+    # fits already produce fp32, so the casts are idempotent guarantees.
     tensors: dict[str, torch.Tensor] = {
-        "node_coords": manifold.node_coords.contiguous().cpu(),
+        "node_coords": manifold.node_coords.contiguous().to(torch.float32).cpu(),
     }
     for idx, sub in manifold.layers.items():
-        tensors[f"layer_{idx}.mean"] = sub.mean.contiguous().cpu()
-        tensors[f"layer_{idx}.basis"] = sub.basis.contiguous().cpu()
-        tensors[f"layer_{idx}.node_params"] = sub.node_params.contiguous().cpu()
-        tensors[f"layer_{idx}.rbf_weights"] = sub.rbf_weights.contiguous().cpu()
-        tensors[f"layer_{idx}.poly_coeffs"] = sub.poly_coeffs.contiguous().cpu()
-        tensors[f"layer_{idx}.coord_offset"] = sub.coord_offset.contiguous().cpu()
-        tensors[f"layer_{idx}.coord_scale"] = sub.coord_scale.contiguous().cpu()
+        tensors[f"layer_{idx}.mean"] = sub.mean.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.basis"] = sub.basis.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.node_params"] = sub.node_params.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.rbf_weights"] = sub.rbf_weights.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.poly_coeffs"] = sub.poly_coeffs.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.coord_offset"] = sub.coord_offset.contiguous().to(torch.float32).cpu()
+        tensors[f"layer_{idx}.coord_scale"] = sub.coord_scale.contiguous().to(torch.float32).cpu()
     save_file(tensors, str(path))
 
     from saklas import __version__ as _saklas_version
